Sweet Alert Added

+++ client/src/components/Post/Post.js
@@ -6,6 +6,7 @@ import Sidebar from "../Sidebar";
 import PostEditModal from "./PostEditModal";
 import PropTypes from "prop-types";
 import { TransitionGroup, CSSTransition } from "react-transition-group";
+import Swal from "sweetalert2";
 
 import { FaTrash } from "react-icons/fa";
 
@@ -32,7 +33,20 @@ export class Post extends Component {
   };
 
   deletePost = (id) => {
-    this.props.deletePost(id);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.props.deletePost(id);
+        Swal.fire("Deleted!", "Your file has been deleted.", "success");
+      }
+    });
   };
 
   render() {
